a fill scratch stage added
    a problem: cfront change declaration of the c functions.
    original declaration:
    	extern "C" void *malloc(unsigned);
    after cfront-1: extern int  *malloc();
    after cfront-3: extern char *malloc();

+++ incl/CC/malloc.h
@@ -5,10 +5,10 @@
 
 extern "C" {
     #ifndef __GNUC__
-	void *malloc(size_t);
+	void *malloc(unsigned);
 	void free(void *);
-	void *realloc(void *, size_t);
-	void *calloc(size_t, size_t);
+	void *realloc(void *, unsigned);
+	void *calloc(unsigned, unsigned);
     #else
 	#include <stdlib.h>
     #endif
